Changed player damage to a float

+++ Project Riptide/Assets/Scripts/Combat/CannonBallBehavior.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class CannonBallBehavior : MonoBehaviour
 {
-    public int damageDealt;
+    public float damageDealt;
     public List<StatusEffect> onhitEffects;
     public GameObject hitbox;
     public string fireSound;
